[#244] feat: move KSingleton and KTransient to core.di
