+++ backend/backend-lwjgl/src/main/java/io/github/darthakiranihil/konna/backend/lwjgl/glfw/KGlfwLwjgl.java
@@ -18,7 +18,7 @@ package io.github.darthakiranihil.konna.backend.lwjgl.glfw;
 
 
 import io.github.darthakiranihil.konna.core.object.KObject;
-import io.github.darthakiranihil.konna.core.object.KSingleton;
+import io.github.darthakiranihil.konna.core.di.KSingleton;
 import io.github.darthakiranihil.konna.test.KExcludeFromGeneratedCoverageReport;
 import io.github.darthakiranihil.konna.libfrontend.glfw.*;
 import org.jspecify.annotations.Nullable;
