+++ backend/backend-spair/src/main/java/io/github/darthakiranihil/konna/backend/spair/imgui/KImGuiSpair.java
@@ -19,7 +19,7 @@ package io.github.darthakiranihil.konna.backend.spair.imgui;
 import imgui.ImGui;
 import imgui.type.*;
 import io.github.darthakiranihil.konna.core.object.KObject;
-import io.github.darthakiranihil.konna.core.object.KSingleton;
+import io.github.darthakiranihil.konna.core.di.KSingleton;
 import io.github.darthakiranihil.konna.core.struct.KVector2f;
 import io.github.darthakiranihil.konna.core.struct.KVector4f;
 import io.github.darthakiranihil.konna.core.struct.ref.*;
