+++ core/src/main/java/io/github/darthakiranihil/konna/core/object/KObjectInstantiationType.java
@@ -16,6 +16,9 @@
 
 package io.github.darthakiranihil.konna.core.object;
 
+import io.github.darthakiranihil.konna.core.di.KSingleton;
+import io.github.darthakiranihil.konna.core.di.KTransient;
+
 /**
  * Enumeration that represents the way {@link KActivator} creating a new object
  * of given class. Normally, it is used only by {@link KActivator}
