+++ implementations/implementations-glfw/src/main/java/io/github/darthakiranihil/konna/core/app/glfw/KGlfwFrame.java
@@ -26,7 +26,7 @@ import io.github.darthakiranihil.konna.core.io.control.KKey;
 import io.github.darthakiranihil.konna.core.io.control.KKeyActionType;
 import io.github.darthakiranihil.konna.core.io.control.KKeyInputData;
 import io.github.darthakiranihil.konna.core.object.KObject;
-import io.github.darthakiranihil.konna.core.object.KSingleton;
+import io.github.darthakiranihil.konna.core.di.KSingleton;
 import io.github.darthakiranihil.konna.core.struct.KSize;
 import io.github.darthakiranihil.konna.libfrontend.glfw.KGlfw;
 import io.github.darthakiranihil.konna.libfrontend.glfw.KGlfwCallbacks;
