+++ implementations/implementations-std-core/src/main/java/io/github/darthakiranihil/konna/core/message/KStandardEventSystem.java
@@ -19,7 +19,7 @@ package io.github.darthakiranihil.konna.core.message;
 import io.github.darthakiranihil.konna.core.except.KNotFoundException;
 import io.github.darthakiranihil.konna.core.log.system.KSystemLogger;
 import io.github.darthakiranihil.konna.core.object.KObject;
-import io.github.darthakiranihil.konna.core.object.KSingleton;
+import io.github.darthakiranihil.konna.core.di.KSingleton;
 import io.github.darthakiranihil.konna.core.object.KTag;
 import io.github.darthakiranihil.konna.core.struct.KStructUtils;
 import org.jspecify.annotations.Nullable;
