+++ implementations/implementations-std-core/src/main/java/io/github/darthakiranihil/konna/core/object/KStandardObjectRegistry.java
@@ -16,6 +16,7 @@
 
 package io.github.darthakiranihil.konna.core.object;
 
+import io.github.darthakiranihil.konna.core.di.KSingleton;
 import io.github.darthakiranihil.konna.core.struct.KStructUtils;
 
 import java.util.*;
