+++ implementations/implementations-std-core/src/main/java/io/github/darthakiranihil/konna/core/util/KHashMapBasedCache.java
@@ -18,7 +18,7 @@ package io.github.darthakiranihil.konna.core.util;
 
 import io.github.darthakiranihil.konna.core.log.system.KSystemLogger;
 import io.github.darthakiranihil.konna.core.object.KObject;
-import io.github.darthakiranihil.konna.core.object.KSingleton;
+import io.github.darthakiranihil.konna.core.di.KSingleton;
 import io.github.darthakiranihil.konna.core.object.KTag;
 import io.github.darthakiranihil.konna.core.struct.KStructUtils;
 import org.jspecify.annotations.Nullable;
