+++ implementations/implementations-std-core/src/test/java/io/github/darthakiranihil/konna/core/object/impl/TestImmortal.java
@@ -17,7 +17,7 @@
 package io.github.darthakiranihil.konna.core.object.impl;
 
 import io.github.darthakiranihil.konna.core.object.KObject;
-import io.github.darthakiranihil.konna.core.object.KSingleton;
+import io.github.darthakiranihil.konna.core.di.KSingleton;
 
 @KSingleton(immortal = true)
 public class TestImmortal extends KObject {
